Update send_backup.rsc
Now the script gets the configuration size and writes it to a file. This size will be used for comparison, so as not to save the RSC file in memory. And a block has been added to check the result of sending an email, if sending an email fails, the configuration size will not be updated.